Moves WE persistence logic into separate component.
